Clarify filter_entry docs
Currently the docs for `filter_entry` state that "Note that if the
iterator has `contents_first` enabled, then this method is no different than calling the standard `Iterator::filter` method, which is slightly wrong, since `Iterator::filter` would take a `Result<&DirEntry, Error>` instead of a `&DirEntry`. The method which would correspond to `filter_entry` is `Iterator::filter_ok`.
I just made the mistake of not reading the `filter_entry` docs
correctly, overlooking that directories which don't satisfy the
predicate are not descended into. Changing `filter_entry` to `filter_ok`
allowed me to use the same predicate and get the behaviour that I
actually wanted.